updates for win64

+++ src/libxdr/xdr_mem.c
@@ -54,7 +54,6 @@
 #include "rpc/xdr.h"
 #endif
 
-
 static bool_t xdrmem_getlong (XDR *, long *);
 static bool_t xdrmem_putlong (XDR *, const long *);
 static bool_t xdrmem_getbytes (XDR *, caddr_t, u_int);
@@ -146,8 +145,7 @@ xdrmem_putbytes (XDR *xdrs, const char *addr, u_int len)
 static u_int
 xdrmem_getpos(const XDR *xdrs)
 {
-
-  return ((u_int)xdrs->x_private - (u_int)xdrs->x_base);
+  return ((uintptr_t)xdrs->x_private - (uintptr_t)xdrs->x_base);
 }
 
 static bool_t
@@ -155,11 +153,10 @@ xdrmem_setpos(register XDR *xdrs, u_int pos)
 {
   register caddr_t newaddr = xdrs->x_base + pos;
   register caddr_t lastaddr = xdrs->x_private + xdrs->x_handy;
-
-  if ((long)newaddr > (long)lastaddr)
+  if ((uintptr_t)newaddr > (uintptr_t)lastaddr)
     return (FALSE);
   xdrs->x_private = newaddr;
-  xdrs->x_handy = (int)lastaddr - (int)newaddr;
+  xdrs->x_handy = (uintptr_t)lastaddr - (uintptr_t)newaddr;
   return (TRUE);
 }
